pyproject.toml
[tool.poetry]
name = "robot_dlya_antonova"
version = "0.1.0"
description = ""
authors = ["Arthur Galimyanov <galart777@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
websockets = "^8.1"
asyncio = "^3.4.3"
nest-asyncio = "^1.5.1"
python-i18n = {extras = ["yaml"], version = "^0.3.9"}
aiounittest = "^1.4.0"
mysql-connector-python = "^8.0.23"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
robot = "robot_dlya_antonova.robot:main"